Create Project
project is created and fields for General Info and
Experience are added with input fields for the
appropriate sections with submit buttons that
will print the values and remove the form fields.
printing the values also reveals an edit button
that will bring the relevant form field back into
view. there is also functionality to add and
delete jobs in the experience section.